# gh-118761: Improve import time of `tomllib`