@vstinner 2025-01-16
cc @encukou who added the module in Python 3.11 (PEP 680).